% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_collection.R
\name{add_collection}
\alias{add_collection}
\title{Add new vectors and/or metadata to the collection}
\usage{
add_collection(
  db,
  vectors = NULL,
  metadatas,
  model = "text-embedding-ada-002",
  url = "https://api.openai.com/v1/embeddings",
  api_key = Sys.getenv("OPENAI_API_KEY")
)
}
\arguments{
\item{db}{The vector database}

\item{vectors}{A list of vectors to add. If NULL, vectors are retrieved using the text field in the metadata.}

\item{metadatas}{A list of metadata corresponding to each vector}
}
\value{
The updated vector database
}
\description{
This function retrieves new vectors from an external source using the provided API key
(if not already provided) and adds them, along with any metadata, to the specified vector database.
If vectors are not provided, the function uses the text field in the metadata to query an API
to retrieve the vectors. The id field in the metadata is created as a reproducible hash of the
embedding vector.
}
\examples{
\dontrun{
db <- add_collection(create_collection(),
                 vectors = list(c(1.2, 2.3, 4.5),
                                   c(6.7, 8.2, 9.2),
                                   c(3.5, 6.7, 8.1),
                                   c(1.8, 7.2, 6.3),
                                   c(5.6, 7.2, 8.9)),
                 metadatas = list(list(text = "This is a document", file = "source1"),
                                  list(text = "This is another document", file = "source2"),
                                  list(text = "This is yet another document", file = "source3"),
                                  list(text = "This is a fourth document", file = "source4"),
                                  list(text = "This is the fifth document", file = "source5"))
                 )
}
}
