% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.R
\name{alpha}
\alias{alpha}
\title{Modify colour transparency}
\usage{
alpha(x, alpha = 0.5)
}
\arguments{
\item{x}{vector of colors specified as hex strings or named R colors.}

\item{alpha}{transparency, number in \code{[0, 1]}; 0 means fully transparent, 1 means fully opaque.}
}
\description{
Modify colour transparency, vectorised in both \code{x} and \code{alpha} (when both are vectors, they must be the same length).
}
\details{
All other operations in chroma do not support alpha channels (because the underlying library, chroma.js, does not output 8 digits hex color specification, which describe r, g, b, and a channels). Therefore, any manipulation of transparency must be done last.
}
\examples{
as.hex("red")
alpha("red")

show_col(c("red", alpha("red")))

# Vectorised in both x and alpha, as long as the lengths are compatible
alpha(c("red", "green", "blue"), 0.5)
alpha("red", c(0.2, 0.5, 0.7))
alpha(c("red", "green", "blue"), c(0.2, 0.5, 0.7))
alpha(c("red", "green", "blue", "purple"), c(0.2, 0.5))
\dontrun{alpha(c("red", "green", "blue", "purple"), c(0.2, 0.5, 0.7))
# = fails because arguments have incompatible lengths (x = 4, alpha = 3)}
# Beware, other functions do not support alpha channels
mix(alpha("red"), alpha("blue"))
mix("red", "blue")
# = both are the same, and fully opaque. Use alpha last
alpha(mix("red", "blue"))
}
