% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_average.R
\name{average}
\alias{average}
\title{Average several colors}
\usage{
average(x, model = "lab")
}
\arguments{
\item{x}{vector of colors specified as hex strings or named R colors.}

\item{model}{color model to compute the averaging in; valid models are \code{rgb}, \code{lrgb}, \code{hsv}, \code{hsl}, \code{hcl}, \code{lab}.}
}
\value{
A vector of colors specified as hex codes
}
\description{
Average several colors
}
\examples{
average(c("red", "blue"))
average(c("red", "blue", "green"))

# show the input colors, then white, then the average
show_ave <- function(x, ...) {show_col(c(x, "white", average(x, ...)))}

show_ave(c("#FEF213", "#146EFD"))
show_ave(hue_colors(5))
show_ave(brewer_colors(5, name="Set3"))
show_ave(cubehelix_colors(5))
# so averaging produces brown, most of the time, but not always:
show_ave(brewer_colors(5, name="Greens"))

# beware, numerical color averaging is not the same as blending pigments
show_ave(c("yellow", "red"))
show_ave(c("blue", "red"))
# but
show_ave(c("yellow", "blue"))
show_ave(c("yellow", "blue"), model="rgb")
show_ave(c("yellow", "blue"), model="hsv")
}
\seealso{
Other color manipulation functions: \code{\link{channel}},
  \code{\link{darken}}, \code{\link{luminance}},
  \code{\link{mix}}, \code{\link{saturate}}
}
