% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_lightness.R
\name{change_lightness}
\alias{brighten}
\alias{brighter}
\alias{change_lightness}
\alias{darken}
\alias{darker}
\alias{lighten}
\alias{lighter}
\title{Change color lightness}
\usage{
darken(x, amount = 1)

darker(x, amount = 1)

brighten(x, amount = 1)

brighter(x, amount = 1)

lighten(x, amount = 1)

lighter(x, amount = 1)
}
\arguments{
\item{x}{vector of colors specified as hex strings or named R colors.}

\item{amount}{numeric amount of change in lightness. Reasonable amounts are 1 to 10. Negative amounts correspond to the opposite operation (\code{darken("red", -1) = brighten("red", 1)}.)}
}
\value{
A vector of colors specified as hex codes
}
\description{
Darken or brighten colors
}
\details{
Colors are converted into L* a* b* space where the L* component is changed. Most colors are (very slightly) affected by the conversion and the change in lightness is therefore not exactly reversible (brightening a darkened color will not get you back to the original one); although, perceptually, the changes should be extremely subtle and only affect very bright colors.
}
\examples{
darken("#7BBBFE")
darken(c("coral1", "#850E5D"))
brighten("darkgreen")

show_col(c("deeppink", darken("deeppink"), brighten("deeppink")),
         c("blue", darken("blue"), brighten("blue")))
show_col(c(brighten("red", 2),
           brighten("red"),
           "red",
           darken("red"),
           darken("red", 2)))

# darken and brighten are opposite operations
darken("red", -1)
brighten("red", 1)

# but not necessarily reversible
col <- "#5eff15"
(new_col <- brighten(darken(col)))
show_col(c(col, new_col))

}

