% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chroma_r.R
\name{chroma_r}
\alias{chroma_r}
\title{Named R Color Specification}
\usage{
chroma_r(x)
}
\arguments{
\item{x}{vector of named R colors (can include colors in already hex representation)}
}
\description{
Convert a named R color as defined in \code{\link[grDevices]{colors}} into its hex representation
}
\examples{
chroma_r("pink")
chroma_r(c("pink", "darkblue"))

show_col(chroma_r(colors()))
 
}

