% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_chroma.R
\name{chroma_scale}
\alias{chroma.colors}
\alias{chroma_colors}
\alias{chroma_map}
\alias{chroma_palette}
\alias{chroma_scale}
\title{Lightness scale and palette}
\usage{
chroma_scale(chroma = c(0, 1), l = 0.5, h = 0, domain = c(0, 1),
  reverse = FALSE)

chroma_map(x, ...)

chroma_palette(...)

chroma_colors(n, ...)

chroma.colors(n, ...)
}
\arguments{
\item{chroma}{chromacity, number in \code{[0, ~1]}; 0 is grey, ~1 is full color.}

\item{l}{lightness, number in \code{[0, 1]}; 0 is darkest, 1 is brightest.}

\item{h}{hue, either an angle around the color wheel, in \code{[0, 360]} (angles outside of the range are rotated back to within \code{[0, 360]}: 380 = 20, -15 = 345, etc.), or a color (hex or named) from which the hue is extracted (by function \code{\link{hue}}).}

\item{domain}{the values between which the scale is computed.}

\item{reverse}{whether to reverse the order of colors along the scale.}

\item{x}{a vector whose values will be coerced to numbers and mapped to colors.}

\item{...}{passed to \code{\link{chroma_scale}}. Note that argument \code{domain} is meaningless in functions other than \code{chroma_scale} and passing it through \code{...} is an error.}

\item{n}{number of colors to extract from the color palette.}
}
\value{
\code{*_scale} and \code{*_palette} return a function.

\code{*_map}, \code{*_colors}, and \code{*.colors} return a vector of hex colors.
}
\description{
Lightness-based color scale and palette in HCL space.
}
\details{
In HCL space, the perceived color (hue) is completely separated from the perceived intensity (chromacity) and lightness of the color. This means that colors of various hues but same chromacity and lightness appear as the exact same grey when converted to greyscale. This makes the HCL space particularly suitable to create good color palettes:
\itemize{
  \item For qualitative palettes (discrete variables): varying \code{h} at constant \code{c} and \code{l} avoids drawing attention to certain hues, as would happen if the same was done in HSV or HSL space. Indeed, some hues are perceived as brighter (yellow, light green, etc.), others as duller/darker (blues, etc.).
  \item For sequential palettes (continuous variables): varying \code{l} (or possibly \code{c}) for a constant \code{h} gives a sense of direction and avoid the many perceptual pitfalls typical of 'rainbow'-like scales.
}

The \code{*_scale} version creates a \strong{function} that takes a single argument (\code{x}: a numeric vector), maps the values to colors along an interpolated gradient and returns those colors as hex codes.

The \code{*_map} version is a shortcut for \code{*_scale(domain=range(x))(x)}: it creates a scale that spans all values in its argument \code{x}, maps the content of \code{x} on that scale and returns the colors. It is particularly useful as the \code{col} argument of plotting functions.

The \code{*_palette} version also creates a \strong{function}, but this one takes an integer (\code{n}) as argument, picks \code{n} colors evenly spaced along the scale and returns them as hex codes.

The \code{*_colors} and \code{*.colors} versions are shortcuts for \code{*_palette()(n)} and directly return \code{n} evenly spaced colors. They are equivalent to built-in functions such as \code{\link[grDevices]{heat.colors}}, \code{\link[grDevices]{topo.colors}}, etc.
}
\examples{
# Define a dark-to-light blue scale
reds <- chroma_scale(h=30)
# and apply it to some data
reds(x=c(0, 0.2, 0.6, 1))

# Define a palette function
# (which works like the actual rainbow() function)
reds_pal <- chroma_palette(h=30)
# and get 10 colors from it
reds_pal(n=10)
show_col(blues_pal(n=10))
# or use the shortcut
show_col(chroma_colors(n=50, h=30))

# Determine hue from a color
blues <- chroma_colors(n=50, h="dodgerblue")
greens <- chroma_colors(n=50, h="green")
yellows <- chroma_colors(n=50, h="gold")
pinks <- chroma_colors(n=50, h="deeppink")
show_col(blues, greens, yellows, pinks)

# Lightness scales are good for continuous variables
# such as the elevation of the Maunga Whau volcano
x <- 10*(1:nrow(volcano))
y <- 10*(1:ncol(volcano))
image(x, y, volcano, col=chroma_colors(100, h="red"))
contour(x, y, volcano, col=alpha("white", 0.5), add=TRUE)

persp(x, y, volcano, theta=50, phi=25, border=alpha("black", 0.3),
      col=chroma_map(persp_facets(volcano), h="red"))

\dontrun{library("rgl")
persp3d(x, y, volcano, aspect=c(1,0.6,0.3), axes=FALSE, box=FALSE,
        col=chroma_map(volcano, h="red"))
play3d(spin3d(axis=c(0, 0, 1), rpm=10), duration=6)
}
# With a limited number of levels, they may also work for discrete variables
attach(iris)
plot(Petal.Length, Sepal.Length, pch=19, col=chroma_map(Species))
legend(1, 8, legend=levels(Species), pch=19,
       col=chroma_colors(n=nlevels(Species)))
# but a hue-based scale is probably more appropriate (see ?hue_map)

}
\seealso{
The \code{\link{hcl}} function, on which this one is based.

Other color scales and palettes: \code{\link{brewer_scale}},
  \code{\link{color_scale}}, \code{\link{hue_scale}},
  \code{\link{inferno_scale}}, \code{\link{light_scale}},
  \code{\link{magma_scale}}, \code{\link{plasma_scale}},
  \code{\link{viridis_scale}}

Other HCL-based scales: \code{\link{hue_scale}},
  \code{\link{light_scale}}
}

