% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_CMClc.R
\name{CMClc}
\alias{CMClc}
\title{Compute the CMC l:c difference between two colors}
\usage{
CMClc(x, y, l = 1, c = 1)
}
\arguments{
\item{x, y}{colors specified as hex strings or named R colors.}

\item{l, c}{weighting factors for lightness and chromacity.}
}
\value{
A number quantifying the difference between x and y (or a vector thereof).
}
\description{
Compute the difference between two colors, using the CMC l:c metric. \url{https://en.wikipedia.org/wiki/Color_difference#CMC_l:c_(1984)}.
}
\examples{
CMClc("pink", "hotpink")
CMClc("pink", "blue")

# NB: his definition of distance is not symetrical
CMClc("pink", "blue")
CMClc("blue", "pink")

# The computation can be vectorised
# For example to find the closest color in an array of possibilities
clrs <- rainbow(20)
show_col("pink", clrs)
d <- CMClc("pink", clrs)
show_col("pink", clrs[which.min(d)])
}
\seealso{
\code{\link{deltaE}} for another (probably more consistent) metric of difference between colors.
}
