% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_cmyk.R
\name{cmyk}
\alias{cmyk}
\title{CMYK color specification}
\usage{
cmyk(c = 0, m = 0.6, y = 0.6, k = 0.4)
}
\arguments{
\item{c, m, y, k}{color components, numeric vectors with values in \code{[0, 1]}.}
}
\value{
A vector of colors specified as hex codes
}
\description{
Create a vector of colors from cyan, magenta, yellow, and black
}
\details{
The first argument can also be a data.frame or a matrix. In that case, its columns are considered as the color components, taken in order, and the other color components arguments are ignored.

When separate arguments are used for the color components and are vectors, values in shorter arguments are recycled to match the length of the longest argument. If the lengths are not compatible, an error is output.

In R (and in chroma) colors are represented internally in the sRGB color space. CMYK is the color model used by printers, which use cyan, magenta, yellow and black ink. The CMYK model and its associated color spaces cannot reproduce all sRGB colors, seen on screens; bright and flashy colors in particular. On the other hand, some CMYK colors are not representable in sRGB; in those cases, they will be converted to the nearest sRGB color. It is therefore possible for two different \code{cmyk} specifications to yield the same sRGB hex code.
}
\examples{
cmyk()
cmyk(0, 0, 0, 0.1)
cmyk(data.frame(c(1, 0.5), c(1, 1), c(1, 1), c(0.5, 0.5)))
cmyk(matrix(c(1, 0.5, 1, 1, 1, 1, 0.5, 0.5), ncol=4))
cmyk(c=0.5, m=1, y=1, k=c(0.25, 0.5, 0.75))

# Color ramp
show_col(cmyk(k=seq(0, 1, length.out=10)))
}
\seealso{
\code{\link{parse_color}} for the general function to parse colors in various specifications (which this function calls internally) and \code{\link{convert_color}} to convert parsed colors to another model.

Other color specifications: 
\code{\link{css}()},
\code{\link{hcl}()},
\code{\link{hex}()},
\code{\link{hsi}()},
\code{\link{hsl}()},
\code{\link{hsv}()},
\code{\link{lab}()},
\code{\link{parse_color}()},
\code{\link{rgb}()},
\code{\link{ryb}()},
\code{\link{temperature}()},
\code{\link{wavelength}()}
}
\concept{color specifications}
