% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{color_temperature}
\alias{color_temperature}
\title{Color Temperature Specification}
\usage{
color_temperature(temperature = 5200)
}
\arguments{
\item{temperature}{light temperature in Kelvin, a numeric vector with values in \code{[1000, 40000]}. 1000 is candle light, 2000 is sunset light, 3200 is a tungsten light bulb, 4000 is a white fluorescent light tube, 5200 is average midday sunlight, 7000 is overcast daylight, 10000 is the color of a clear blue sky.}
}
\value{
A vector of colors specified as hex codes
}
\description{
Converts color temperature in Kelvin to a color vector.
}
\examples{
color_temperature()
color_temperature(c(2000, 5200, 7000))

# display range
temp <- seq(1000, 20000, length.out=20)
plot(x=temp, y=rep(1, 20), col=color_temperature(temp), pch=15, cex=3)

}
\seealso{
Other color specifications: \code{\link{cmyk}},
  \code{\link{css_color}}, \code{\link{hcl}},
  \code{\link{hsl}}, \code{\link{hsv}}, \code{\link{lab}},
  \code{\link{parse_color}}, \code{\link{rgb}}
}

