% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_complement.R
\name{complement}
\alias{complement}
\title{Compute the complement color}
\usage{
complement(x, mode = "ryb")
}
\arguments{
\item{x}{vector of colors specified as hex strings or named R colors.}

\item{mode}{color mode to use, either "ryb" which yields the natural, paint-like complement, or "rgb"}
}
\value{
A vector of colors specified as hex codes
}
\description{
Compute the complement color
}
\examples{
# paint-like complements
a <- c("red", "yellow", "blue")
b <- c("green", "purple", "orange")
w <- c("white", "white", "white")
show_col(a, complement(a), b, w, b, complement(b), a)

# scren-like complements
show_col(a, complement(a, mode="ryb"))
show_col(a, complement(a, mode="rgb"))
x <- c("red", "green", "blue", "cyan", "magenta", "yellow")
show_col(x, complement(x, mode="rgb"))
}
