% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrast.R
\name{contrast}
\alias{contrast}
\title{Compute the contrast between two colors}
\usage{
contrast(x, y)
}
\arguments{
\item{x, y}{colors specified as hex strings or named R colors.}
}
\value{
A number giving the contrast between x and y (or a vector thereof).
}
\description{
Compute the contrast between two colors
}
\examples{
contrast("pink", "hotpink")
contrast("pink", "blue")

# The computation can be vectorised
contrast("pink", rainbow(10))

}

