% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_collection.R
\name{create_collection}
\alias{create_collection}
\title{Create a new vector database}
\usage{
create_collection()
}
\value{
A new vector database
}
\description{
This function creates a new in-memory vector database.
}
\examples{
\dontrun{
db <- create_collection()
}
}
