% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_css.R
\name{css}
\alias{css}
\title{CSS color specification}
\usage{
css(x)
}
\arguments{
\item{x}{string representing a color in CSS.}
}
\value{
A vector of colors specified as hex codes
}
\description{
Converts colors specified as CSS strings into R colors. This includes all colors in the X11 specification of the W3C \url{https://www.w3.org/TR/css-color-3/#svg-color} as well as \code{rgb()} and \code{hsl()} constructs (but, in this case, it is probably easier to use the functions \code{\link{rgb}}) and \code{\link{hsl}} directly).
}
\examples{
css("teal")
css(c("teal", "blanchedalmond"))
css("rgb(255, 100, 100)")
css("hsl(200, 50\%, 50\%)")
}
\seealso{
\code{\link{parse_color}} for the general function to parse colors in various specifications (which this function calls internally) and \code{\link{convert_color}} to convert parsed colors to another model.

Other color specifications: 
\code{\link{cmyk}()},
\code{\link{hcl}()},
\code{\link{hex}()},
\code{\link{hsi}()},
\code{\link{hsl}()},
\code{\link{hsv}()},
\code{\link{lab}()},
\code{\link{parse_color}()},
\code{\link{rgb}()},
\code{\link{ryb}()},
\code{\link{temperature}()},
\code{\link{wavelength}()}
}
\concept{color specifications}
