% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_css.R
\name{css_color}
\alias{css_color}
\title{CSS Color Specification}
\usage{
css_color(x)
}
\arguments{
\item{x}{string representing a color in CSS.}
}
\value{
A vector of colors specified as hex codes
}
\description{
Converts colors specified as CSS strings into R colors. This includes all colors in the X11 specification of the W3C \code{<http://www.w3.org/TR/css3-color/#svg-color>} as well as \code{rgb()} and \code{hsl()} constructs (but, in this case, it is probably easier to use the functions \code{\link[grDevices]{rgb}}) and \code{\link{hsl}} directly).
}
\examples{
css_color("teal")
css_color(c("teal", "blanchedalmond"))
css_color("rgb(255, 100, 100)")
css_color("hsl(200, 50\%, 50\%)")

}
\seealso{
Other color specifications: \code{\link{cmyk}},
  \code{\link{color_temperature}}, \code{\link{hcl}},
  \code{\link{hsl}}, \code{\link{hsv}}, \code{\link{lab}},
  \code{\link{parse_color}}, \code{\link{rgb}}
}

