% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_css.R
\docType{data}
\name{css_colors}
\alias{css_colors}
\title{Named CSS Colors}
\format{A data.frame with 148 rows and 2 variables:
\describe{
  \item{name}{the name of the color in the CSS specification.}
  \item{hex}{the corresponding hex code.}
}}
\source{
\url{https://www.w3.org/TR/css-color-4/#named-colors}
}
\usage{
css_colors
}
\description{
Named CSS Colors
}
\seealso{
\code{\link{css}} or \code{\link{hex}} to parse such colors into their hex codes.
}
\keyword{datasets}
