% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manip_darken_brighten.R
\name{darken}
\alias{darken}
\alias{darker}
\alias{brighten}
\alias{brighter}
\alias{lighten}
\alias{lighter}
\title{Change color lightness}
\usage{
darken(x, amount = 1)

darker(x, amount = 1)

brighten(x, amount = 1)

brighter(x, amount = 1)

lighten(x, amount = 1)

lighter(x, amount = 1)
}
\arguments{
\item{x}{vector of colors specified as hex strings or named R colors.}

\item{amount}{numeric amount of change in lightness. Reasonable amounts are 1 to 10. Negative amounts correspond to the opposite operation (\code{darken("red", -1) = brighten("red", 1)}.)}
}
\value{
A vector of colors specified as hex codes
}
\description{
Darken or brighten colors
}
\details{
Colors are converted into L*a*b* space where the L* component is changed. Most colors are (very slightly) affected by the conversion and the change in lightness is therefore not exactly reversible (brightening a darkened color will not get you back to the original one); although, perceptually, the changes should be extremely subtle and only affect very bright colors.
}
\examples{
darken("#7BBBFE")
darken(c("coral1", "#850E5D"))
brighten("darkgreen")

show_col(c("hotpink", darken("hotpink"), brighten("hotpink")),
         c("lightskyblue", darken("lightskyblue"), brighten("lightskyblue")))
show_col(c(brighten("salmon3", 2),
           brighten("salmon3"),
           "salmon3",
           darken("salmon3"),
           darken("salmon3", 2)))

# darken() and brighten() are opposite operations
darken("red", -1)
brighten("red", 1)

# But they are not necessarily exactly reversible
col <- "#5eff15"
(new_col <- brighten(darken(col)))
show_col(c(col, new_col))

}
\seealso{
Other color manipulation functions: \code{\link{channel}},
  \code{\link{luminance}}, \code{\link{mix}},
  \code{\link{saturate}}
}
