% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_delta_e.R
\name{delta_e}
\alias{delta_e}
\title{Compute the difference between two colors}
\usage{
delta_e(x, y)
}
\arguments{
\item{x, y}{colors specified as hex strings or named R colors.}
}
\value{
A number quantifying the difference between x and y (or a vector thereof). A difference < 1 is imperceptible to the human eye.
}
\description{
Compute the difference between two colors, using the CIE Delta-E 2000 formula. \url{https://en.wikipedia.org/wiki/Color_difference#CIEDE2000}.
}
\examples{
delta_e("pink", "hotpink")
delta_e("pink", "blue")

# The computation can be vectorised
# For example to find the closest color in an array of possibilities
clrs <- rainbow(20)
show_col("pink", clrs)
d <- delta_e("pink", clrs)
show_col("pink", clrs[which.min(d)])

}
\references{
Sharma, Gaurav; Wu, Wencheng; Dalal, Edul N. (2005). "The CIEDE2000 color-difference formula: Implementation notes, supplementary test data, and mathematical observations" (PDF). Color Research & Applications (Wiley Interscience) 30 (1): 21–30. doi:10.1002/col.20070
}
