% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_etopo.R
\docType{data}
\name{etopo}
\alias{etopo}
\title{Colors from ETOPO1}
\format{
A data.frame with 43 rows and 2 variables:
\describe{
  \item{altitude}{the altitude.}
  \item{color}{the hex code of the correpsonding color.}
}
}
\source{
\code{http://soliton.vm.bytemark.co.uk/pub/cpt-city/ngdc/tn/ETOPO1.png.index.html}
}
\usage{
etopo
}
\description{
Colors from ETOPO1
}
\seealso{
\code{\link{etopo_scale}} to use the palette.
}
\keyword{datasets}
