% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_hex.R
\name{hex}
\alias{hex}
\title{HEX Color Specification}
\usage{
hex(x)
}
\arguments{
\item{x}{hexadecimal string specifying a color.}
}
\value{
A vector of colors specified as hex codes
}
\description{
Converts colors specified as hexadecimal strings into R colors. This includes the usual format for R colors but also the shorter three letters format. In addition the \code{#} sign can be omitted.
}
\details{
Hexadecimal strings (hex for short) are how colors are usually represented in R. This function just brings a little more versatility in how those can be specified. Note, however, that R can use the 8 letter model to represent color with a transparency (alpha) component and that this function does not support it (and just removes the transparency).
}
\examples{
hex("#ff3399")
hex("ff3399")
hex("ff3399")
hex("F39")

(half_transparent <- alpha("#ff3399", 0.5))
hex(half_transparent)
# the transparency portion (last two positions: '80') is removed

}
\seealso{
\code{\link{parse_color}} for the general function to parse colors in various specifications (which this function calls internally) and \code{\link{convert_color}} to convert parsed colors to another model.

Other color specifications: \code{\link{cmyk}},
  \code{\link{css}}, \code{\link{hcl}}, \code{\link{hsi}},
  \code{\link{hsl}}, \code{\link{hsv}}, \code{\link{lab}},
  \code{\link{parse_color}}, \code{\link{rgb}},
  \code{\link{temperature}}
}

