% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_hsi.R
\name{hsi}
\alias{hsi}
\title{HSI Color Specification}
\usage{
hsi(h = 0, s = 0.5, i = 0.5)
}
\arguments{
\item{h}{hue, either an angle around the color wheel, in [0,360] (angles outside of the range are rotated back to within [0, 360]: 380 = 20, -15 = 345, etc.), or a color (hex or named) from which the hue is extracted (by function \code{\link{hue}}).}

\item{s}{saturation, number \code{[0,1]}; 0 is grey, 1 is full color.}

\item{i}{intensity, number in \code{[0,2]}; 0 is black, 2 is white.}
}
\value{
A vector of colors specified as hex codes
}
\description{
Create a vector of colors from hue, saturation, and intensity.
}
\details{
The first argument can also be a data.frame or a matrix. In that case, its columns are considered as the color components, taken in order, and the other color components arguments are ignored.

When separate arguments are used for the color components and are vectors, values in shorter arguments are recycled to match the length of the longest argument. If the lengths are not compatible, an error is output.

The HSI color model tentatively separates color (hue), color intensity (saturation), and color intensity, which helps with the creation of color palettes compared to RGB. However, while the color components are separated numerically, some confusion remains in the way colors are perceived by the human eye/brain. Indeed, even at constant \code{s} and \code{i} some hues are perceived as brighter (yellow and green for example) and therefore draw the viewer's attention. This is one of the common problems with 'rainbow'-like color scales (which are constructed in HSL or HSV space). See the function \code{\link{luminance}} for a numerical demonstration. For the creation of color palettes, the HCL space (function \code{\link{hcl}}) should be preferred.
}
\examples{
hsi()
hsi(330, 1, 0.6)
hsi(data.frame(c(330, 340), c(1,1), c(0.6, 0.4)))
hsi(matrix(c(330, 340, 1, 1, 0.6, 0.4), ncol=3))
hsi(c(330, 340), 1, 0.6)

# Color ramps
ramp <- seq(0, 1, length.out=10)
show_col(hsi(h=ramp*360), hsi(s=ramp), hsi(i=ramp*2))
}
\seealso{
\code{\link{parse_color}} for the general function to parse colors in various specifications (which this function calls internally) and \code{\link{convert_color}} to convert parsed colors to another model.

Other color specifications: \code{\link{cmyk}},
  \code{\link{css}}, \code{\link{hcl}}, \code{\link{hex}},
  \code{\link{hsl}}, \code{\link{hsv}}, \code{\link{lab}},
  \code{\link{parse_color}}, \code{\link{rgb}},
  \code{\link{temperature}}, \code{\link{wavelength}}
}
