% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsl.R
\name{hsl}
\alias{hsl}
\title{HSL Color Specification}
\usage{
hsl(h = 0, s = 0.5, l = 0.5)
}
\arguments{
\item{h}{hue, numeric vector with values in \code{[0, 360]}. Alternatively, can be a data.frame or a matrix whose first three columns are considered as arguments; in that case, the rest of the function arguments are ignored.}

\item{s}{saturation, numeric vector with values in \code{[0, 1]}}

\item{l}{lightness, same conventions as \code{s}}
}
\value{
A vector of colors specified as hex codes
}
\description{
Create a vector of colors from hue, saturation and lightness.
}
\details{
When the arguments are used separately and are vectors, values in shorter arguments are recycled to match the length of the longest argument. If the lengths are not compatible, an error is output.
}
\examples{
hsl()
hsl(330, 1, 0.6)
hsl(data.frame(c(330, 340), c(1,1), c(0.6, 0.4)))
hsl(matrix(c(330, 340, 1, 1, 0.6, 0.4), ncol=3))
hsl(c(330, 340), 1, 0.6)

show_col(hsl(h=seq(0, 360, length.out=10)))

}

