% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_hex.R
\name{in_hex}
\alias{in_hex}
\title{Force HEX representation}
\usage{
in_hex(x)
}
\arguments{
\item{x}{vector of colors specified as hex strings or named R colors.}
}
\value{
A vector of colors specified as hex codes only
}
\description{
Convert R colors, including named ones as defined in the function \code{\link[grDevices]{colors}}, into their hex representation
}
\examples{
in_hex("pink")
in_hex(c("pink", "#348A31", "darkblue"))

show_col(in_hex(colors()))
 
}

