% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_vectorDB.R
\name{is_vectorDB}
\alias{is_vectorDB}
\title{Check if an object is a vectorDB}
\usage{
is_vectorDB(x)
}
\arguments{
\item{x}{The object to check}
}
\value{
TRUE if the object is a vectorDB, FALSE otherwise
}
\description{
This function checks if an object is a vectorDB by checking its class, structure, and contents.
}
\examples{
\dontrun{
is_vectorDB(db)
}
}
