% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_magma.R
\docType{data}
\name{magma}
\alias{magma}
\title{Magma colors}
\format{
A vector of length 256 containing hex values.
}
\source{
\url{https://bids.github.io/colormap/} for the concept and \url{https://github.com/BIDS/colormap/blob/master/colormaps.py} for the data.
}
\usage{
magma
}
\description{
The \code{magma} color palette, by Nathaniel J. Smith and Stefan van der Walt (CC0 license).
}
\seealso{
Other mpl palettes: 
\code{\link{inferno}},
\code{\link{plasma}},
\code{\link{viridis}}
}
\concept{mpl palettes}
\keyword{datasets}
