% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_objects.R
\docType{data}
\name{maunga}
\alias{maunga}
\alias{maungaxyz}
\title{Topographic information on Auckland's Maunga Whau volcano}
\format{\code{maunga} is a list with two coordinate vectors and a matrix of corresponding values, suitable for \code{\link[graphics]{contour}}, \code{\link[graphics]{persp}} etc. \code{maungaxyz} contains the same data but formated as a xyz a data.frame with 5307 rows and 3 columns. Both have components:
\describe{
  \item{x,y}{x and y coordinates, in m.}
  \item{z}{altitudes, in m.}
}}
\usage{
maunga

maungaxyz
}
\description{
Maunga Whau (Mt Eden) is one of about 50 volcanos in the Auckland volcanic field. This data set gives topographic information for Maunga Whau on a 10m by 10m grid.
}
\seealso{
\code{\link[datasets]{volcano}} for the original data.
}
\keyword{datasets}
