% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_handling.R
\name{na_insert}
\alias{na_insert}
\title{Insert NAs into a vector based on another one}
\usage{
na_insert(x, from)
}
\arguments{
\item{x}{vector without NAs}

\item{from}{vector from which the NAs must be inserted in x}
}
\description{
Insert NAs into a vector based on another one
}
\examples{
x <- c(1, NA, 2)
y <- na.omit(x)
na_insert(y^2, from=x)
}
