% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_color.R
\name{parse_color}
\alias{parse_color}
\title{Parse colors specified in a given model}
\usage{
parse_color(x, model)
}
\arguments{
\item{x}{a matrix or data.frame whose column specify color channels.}

\item{model}{string defining the color model; valid models are \code{rgb}, \code{hsv}, \code{hsl}, \code{hcl}, \code{lch}, \code{lab}, \code{cmyk}}
}
\value{
A vector of colors specified as hex codes
}
\description{
Parse colors specified in a given model
}
\examples{
parse_color(data.frame(h=c(0, 120, 240), s=0.5, v=0.7), "hsv")
parse_color(data.frame(r=c(255, 0, 0),
                       g=c(0, 255, 0),
                       b=c(0, 0, 255)), "rgb")

}
\seealso{
Other color specifications: \code{\link{cmyk}},
  \code{\link{color_temperature}}, \code{\link{css_color}},
  \code{\link{hcl}}, \code{\link{hsl}}, \code{\link{hsv}},
  \code{\link{lab}}
}

