% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persp_facets.R
\name{persp_facets}
\alias{persp_facets}
\title{Facets values in a persp plot}
\usage{
persp_facets(z)
}
\arguments{
\item{z}{a numeric matrix to be used as the z argument of \code{link[graphics]{persp}}}
}
\value{
A matrix of dimensions \code{[nrow(z)-1, ncol(z)-1]} with the values at the center of each facet
}
\description{
Compute the value at the center of facets in a perspective plot of a matrix
}
\examples{
z <- matrix(1:12, nrow=4)
persp_facets(z)

}
\seealso{
\code{\link{color_scale}} for an example of usage.
}
