% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_plasma.R
\docType{data}
\name{plasma}
\alias{plasma}
\title{Magma colors}
\format{A vector of length 256 containing hex values.}
\source{
\url{https://bids.github.io/colormap/} for the concept and \url{https://github.com/BIDS/colormap/blob/master/colormaps.py} for the data.
}
\usage{
plasma
}
\description{
The \code{plasma} color palette, by Nathaniel J. Smith and Stefan van der Walt (CC0 license).
}
\seealso{
Other mpl palettes: \code{\link{inferno}},
  \code{\link{magma}}, \code{\link{viridis}}
}
\keyword{datasets}

