% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_plasma.R
\name{plasma_scale}
\alias{plasma_scale}
\alias{plasma_scale}
\alias{plasma_map}
\alias{plasma_scale}
\alias{plasma_palette}
\alias{plasma_scale}
\alias{plasma_colors}
\alias{plasma_scale}
\alias{plasma.colors}
\title{Magma color scale and palette}
\usage{
plasma_scale(domain = c(0, 1), reverse = FALSE)

plasma_map(x, ...)

plasma_palette(...)

plasma_colors(n, ...)

plasma.colors(n, ...)
}
\arguments{
\item{domain}{the values between which the scale is computed.}

\item{reverse}{whether to reverse the order of colors along the scale.}

\item{x}{a vector whose values will be coerced to numbers and mapped to colors.}

\item{...}{passed to \code{\link{plasma_scale}}. Note that argument \code{domain} is meaningless in functions other than \code{plasma_scale} and passing it through \code{...} is an error.}

\item{n}{number of colors to extract from the color palette.}
}
\value{
\code{*_scale} and \code{*_palette} return a function.

\code{*_map}, \code{*_colors}, and \code{*.colors} return a vector of hex colors.
}
\description{
The \code{plasma} color palette, by Nathaniel J. Smith and Stefan van der Walt (CC0 license).
}
\details{
The \code{*_scale} version creates a \strong{function} that takes a single argument (\code{x}: a numeric vector), maps the values to colors along an interpolated gradient and returns those colors as hex codes.

The \code{*_map} version is a shortcut for \code{*_scale(domain=range(x))(x)}: it creates a scale that spans all values in its argument \code{x}, maps the content of \code{x} on that scale and returns the colors. It is particularly useful as the \code{col} argument of plotting functions.

The \code{*_palette} version also creates a \strong{function}, but this one takes an integer (\code{n}) as argument, picks \code{n} colors evenly spaced along the scale and returns them as hex codes.

The \code{*_colors} and \code{*.colors} versions are shortcuts for \code{*_palette()(n)} and directly return \code{n} evenly spaced colors. They are equivalent to built-in functions such as \code{\link[grDevices]{heat.colors}}, \code{\link[grDevices]{topo.colors}}, etc.
}
\examples{
# Get a few colors along the palette
show_col(
  plasma_palette()(20),
  plasma_colors(50),
  plasma_colors(20, reverse=TRUE)
)

# Plot the Maunga Whau volcano elevation map
image(maunga, col=plasma_colors(100))
contour(maunga, col=alpha("black", 0.2), add=TRUE)

persp(maunga, theta=50, phi=25, border=alpha("black", 0.3),
      col=plasma_map(persp_facets(maunga$z)))

\dontrun{library("rgl")
persp3d(maunga, aspect=c(1,0.6,0.3), axes=FALSE, box=FALSE,
        col=plasma_map(maunga$z))
play3d(spin3d(axis=c(0, 0, 1), rpm=10), duration=6)
}
# With a limited number of levels, plasma can also serve as a discrete
# color scale
attach(iris)
plot(Petal.Length, Sepal.Length, pch=19, col=plasma_map(Species))
legend(1, 8, legend=levels(Species), pch=19,
       col=plasma_colors(n=nlevels(Species)))

}
\seealso{
\code{\link{plasma}} for the colors in the palette.

Other color scales and palettes: \code{\link{brewer_scale}},
  \code{\link{chroma_scale}},
  \code{\link{cubehelix_scale}}, \code{\link{hue_scale}},
  \code{\link{inferno_scale}}, \code{\link{interp_scale}},
  \code{\link{light_scale}}, \code{\link{magma_scale}},
  \code{\link{viridis_scale}}
}
