% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_collection.R
\name{query_collection}
\alias{query_collection}
\title{Query a vector database}
\usage{
query_collection(
  db,
  filter = NULL,
  query_embeddings = NULL,
  top_n = NULL,
  type = "cosine"
)
}
\arguments{
\item{db}{The vector database}

\item{filter}{An expression string that represents the filter to be applied on metadata}

\item{query_embeddings}{A vector or other vector database to compute similarity against}

\item{top_n}{Integer, indicating the number of top matches to return}

\item{type}{The function to use to compute the similarities between the vectors}
}
\value{
A matrix of similarity scores or a data.table of top matches
}
\description{
This function filters a vector database based on provided metadata constraints and/or computes
similarity of the remaining vectors to a provided embedding, a text string, or another database.
}
\examples{
\dontrun{
db <- add_collection(create_collection(),
                 metadatas = list(list(text = "This is a document", file = "source1"),
                                  list(text = "This is another document", file = "source2"),
                                  list(text = "This is yet another document", file = "source3"),
                                  list(text = "This is a fourth document", file = "source4"),
                                  list(text = "This is the fifth document", file = "source5"))
                 )
similarity_internal <- query_collection(db,
                               query_embeddings = db,
                               top_n = 2)

similarity <- query_collection(db,
                               filter = "file == 'source1'",
                               query_embeddings = c(1.2, 2.3, 4.5),
                               top_n = 5)
}
}
