% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_color.R
\name{random_color}
\alias{random_color}
\title{Generate random color(s)}
\usage{
random_color(n = 1)
}
\arguments{
\item{n}{number of colors to generate}
}
\value{
A vector of colors specified as hex codes
}
\description{
Generate random color(s)
}
\examples{
random_color()
random_color(n=10)
show_col(random_color(n=10), random_color(n=10), random_color(n=10))
# Art!

}

