% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_colors.R
\name{random_colors}
\alias{random_colors}
\alias{rcolors}
\title{Generate random color(s)}
\usage{
random_colors(n = 1)

rcolors(n = 1)
}
\arguments{
\item{n}{number of colors to generate}
}
\value{
A vector of colors specified as hex codes
}
\description{
Generate random color(s)
}
\examples{
random_colors()
random_colors(n=10)
show_col(random_colors(n=10), random_colors(n=10), random_colors(n=10))
# Art!

}
