% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_embeddings.R
\name{retrieve_vectors}
\alias{retrieve_vectors}
\title{Retrieve embeddings from OpenAI}
\usage{
retrieve_vectors(
  inputs,
  model = "text-embedding-ada-002",
  url = "https://api.openai.com/v1/embeddings",
  api_key = Sys.getenv("OPENAI_API_KEY")
)
}
\arguments{
\item{inputs}{A list of text inputs}

\item{model}{The OpenAI embedding model to use}

\item{url}{The OpenAI api url to use}

\item{api_key}{The OpenAI api key to use}
}
\value{
A list of numeric vectors
}
\description{
This function retrieves embeddings for a list of text inputs using OpenAI's API.
}
\examples{
\dontrun{
embeddings <- retrieve_vectors(c("This is a document", "This is another document"))
}
}
