% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgb.R
\name{rgb}
\alias{rgb}
\title{RGB Color Specification}
\usage{
rgb(red = 0, green = 0, blue = 0, alpha = NULL, names = NULL,
  maxColorValue = 1)
}
\arguments{
\item{red, green, blue}{color channels, numeric vectors with values in \code{[0, maxColorValue]}.}

\item{alpha}{transparency, numeric vector with values in \code{[0, maxColorValue]}; 0 means fully transparent, \code{maxColorValue} means fully opaque.}

\item{names}{character vector. The names for the resulting vector.}

\item{maxColorValue}{number giving the maximum of the color values range; typically 1 or 255.}
}
\value{
A vector of colors specified as hex codes
}
\description{
Create a vector of colors from red, green, and blue.
}
\details{
The first argument can also be a data.frame or a matrix. In that case, its first three columns are considered as the three color components and the rest of the function arguments are ignored.

When the arguments are used separately and are vectors, values in shorter arguments are recycled to match the length of the longest argument. If the lengths are not compatible, an error is output.
}
\examples{
rgb()
rgb(1, 0, 0)
rgb(255, 0, 0, maxColorValue=255)

red <- rgb(1, 0, 0)
transparent_red <- rgb(1, 0, 0, alpha=0.7)
show_col(c(red, transparent_red))
    
ramp <- seq(0, 1, length.out=10)
rgb(red=ramp, green=0, blue=0, names=paste("red", 1:10, sep="."))

show_col(
  rgb(r=ramp),
  rgb(g=ramp),
  rgb(b=ramp)
)

}
\seealso{
Other color specifications: \code{\link{cmyk}},
  \code{\link{color_temperature}}, \code{\link{css_color}},
  \code{\link{hcl}}, \code{\link{hsl}}, \code{\link{hsv}},
  \code{\link{lab}}, \code{\link{parse_color}}
}

