% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_xy_map.R
\name{scale_xy_map}
\alias{scale_xy_map}
\title{ggplot2 x and y scales definitions for maps}
\usage{
scale_xy_map(...)
}
\arguments{
\item{...}{passed to \code{\link[ggplot2]{scale_x_continuous}} \strong{and} \code{\link[ggplot2]{scale_y_continuous}}}
}
\description{
Define nice looking x and y axes conventions for maps plotted with ggplot2
}
\details{
This sets \code{expand=c(0,0)} to maximum map extent, map-py axes titles and labels, and removes minor breaks to be consistent among \code{\link[ggplot2]{coord_quickmap}} and \code{\link[ggplot2]{coord_map}}.
}
\examples{
\dontrun{
library("ggplot2")
ggplot(thaixyz) + coord_quickmap() +
  geom_raster(aes(x=x, y=y, fill=z)) +
  scale_fill_wikitopo() +
  scale_xy_map()
 }
}
