% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_objects.R
\docType{data}
\name{scicol}
\alias{scicol}
\title{Scientific Colour Maps}
\format{
An object of class \code{list} of length 24.
}
\source{
https://www.fabiocrameri.ch/colourmaps/
}
\usage{
scicol
}
\description{
Scientific Colour Maps
}
\keyword{datasets}
