% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_objects.R
\docType{data}
\name{scicol_info}
\alias{scicol_info}
\title{Scientific Colour Maps Info}
\format{
An object of class \code{data.frame} with 24 rows and 2 columns.
}
\source{
https://www.fabiocrameri.ch/colourmaps/
}
\usage{
scicol_info
}
\description{
Scientific Colour Maps Info
}
\keyword{datasets}
