% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidelegend.R
\name{sidelegend}
\alias{sidelegend}
\alias{sidemargin}
\title{Put the legend on the side of a plot}
\usage{
sidelegend(...)

sidemargin()
}
\arguments{
\item{...}{passed to \code{\link[graphics]{legend}}.}
}
\value{
\code{sidemargin} returns the previous \code{\link[graphics]{par}} settings, so that they can be reset for futher plots.
\code{sidelegend} returns the same value as \code{\link[graphics]{legend}}.
}
\description{
\code{sidemargin} prepares the margin, \code{sidelegend} puts the legend in it.
}
\examples{
# Make a basic plot
attach(iris)
plot(Petal.Length, Petal.Width, col=Species, pch=19)
legend(1, 2, legend=levels(Species), col=1:nlevels(Species), pch=19)

# The same with the legend on the side
pars <- sidemargin()
plot(Petal.Length, Petal.Width, col=Species, pch=19)
sidelegend(legend=levels(Species), col=1:nlevels(Species), pch=19)
par(pars) # reset graphical parameters
}
