% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_temperature.R
\name{temperature}
\alias{temperature}
\title{Color Temperature Specification}
\usage{
temperature(x = 5200)
}
\arguments{
\item{x}{light temperature in Kelvin, a numeric vector with values in \code{[1000, 40000]}. 1000 is candle light, 2000 is sunset light, 3200 is a tungsten light bulb, 4000 is a white fluorescent light tube, 5200 is average midday sunlight, 7000 is overcast daylight, 10000 is the color of a clear blue sky.}
}
\value{
A vector of colors specified as hex codes
}
\description{
Converts color temperature in Kelvin to a color vector.
}
\examples{
temperature()
temperature(c(2000, 5200, 7000))

# Display a large range of color temperatures
x <- seq(1000, 15000, length.out=200)
notable_temperatures <- c(1000, 2000, 3200, 4000, 5200, 7000, 10000)
plot(x=x, y=rep(0, 200), col=temperature(x), pch="|", cex=5, xaxt='n')
axis(side=1, at=notable_temperatures, labels=TRUE, las=2)
abline(v=notable_temperatures, lty="dotted")
}
\seealso{
\code{\link{parse_color}} for the general function to parse colors in various specifications (which this function calls internally) and \code{\link{convert_color}} to convert parsed colors to another model.

Other color specifications: \code{\link{cmyk}},
  \code{\link{css}}, \code{\link{hcl}}, \code{\link{hex}},
  \code{\link{hsi}}, \code{\link{hsl}}, \code{\link{hsv}},
  \code{\link{lab}}, \code{\link{parse_color}},
  \code{\link{rgb}}, \code{\link{ryb}},
  \code{\link{wavelength}}
}
\concept{color specifications}
