% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_turbo.R
\docType{data}
\name{turbo}
\alias{turbo}
\title{Turbo colors}
\format{
A vector of length 256 containing hex values.
}
\source{
\url{https://ai.googleblog.com/2019/08/turbo-improved-rainbow-colormap-for.html}
}
\usage{
turbo
}
\description{
The \code{turbo} color palette, by Anton Mikhailov (Apache-2.0 license).
}
\concept{palettes}
\keyword{datasets}
