% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v8_chroma_context.R
\name{v8_chroma_context}
\alias{v8_chroma_context}
\title{Create a chroma-aware javascript interpreting context}
\usage{
v8_chroma_context()
}
\description{
Use the \code{\link[V8]{v8}} function of package \code{V8} to create the context and evaluate the code of chroma.js, thus making all functions available.
}
\examples{
ct <- v8_chroma_context()
ct$eval("chroma('pink').darken().saturate(2).hex()")

}
\seealso{
\code{\link[V8]{v8}}
}

