% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_viridis.R
\name{viridis_scale}
\alias{viridis_scale}
\alias{viridis_map}
\alias{viridis_palette}
\alias{viridis_colors}
\title{Viridis color scale and palette}
\usage{
viridis_scale(domain = c(0, 1), reverse = FALSE)

viridis_map(x, ...)

viridis_palette(...)

viridis_colors(n, ...)
}
\arguments{
\item{domain}{vector of two values between which the scale is computed.}

\item{reverse}{whether to reverse the order of colors along the scale.}

\item{x}{a vector whose values will be coerced to numbers and mapped to colors.}

\item{...}{passed to \code{\link{viridis_scale}}. Note that argument \code{domain} is meaningless in functions other than \code{viridis_scale} and passing it through \code{...} is an error.}

\item{n}{number of colors to extract from the color palette.}
}
\value{
\code{*_scale} returns a \strong{function}. This function takes a single argument (\code{x}: a numeric vector), maps its values to colors, and returns thee colors as hex codes.

\code{*_map} is a shortcut for \code{*_scale(domain=range(x))(x)}: it creates a scale that spans the range of values in argument \code{x}, maps the content of \code{x} on that scale, and returns the \strong{colors}.

\code{*_palette} returns a \strong{function}. This function takes an integer (\code{n}) as argument, picks \code{n} colors evenly spaced along the scale, and returns them as hex codes.

\code{*_colors} is a shortcut for \code{*_palette()(n)} and directly returns \code{n} evenly spaced \strong{colors}. It is equivalent to built-in functions such as \code{\link[grDevices]{heat.colors}}, \code{\link[grDevices]{topo.colors}}, etc.

\code{scale_*} return a \strong{ggplot2 scale}, either discrete (similar to \code{\link[ggplot2]{scale_color_discrete}}) or continuous (similar to \code{\link[ggplot2]{scale_color_continuous}}).
}
\description{
The \code{viridis} color palette, by Nathaniel J. Smith, Stefan van der Walt, and Eric Firing (CC0 license).
}
\examples{
# Get a few colors along the palette
show_col(
  viridis_palette()(20),
  viridis_colors(50),
  viridis_colors(20, reverse=TRUE)
)

# Plot the Maunga Whau volcano elevation map
image(maunga, col=viridis_colors(100))
contour(maunga, col=alpha("black", 0.2), add=TRUE)

persp(maunga, theta=50, phi=25, scale=FALSE, expand=2,
      border=alpha("black", 0.4),
      col=viridis_map(persp_facets(maunga$z)))

\dontrun{
library("rgl")
persp3d(maunga, aspect=c(1,0.7,0.2), axes=FALSE, box=FALSE,
        col=viridis_map(maunga$z))
play3d(spin3d(axis=c(0, 0, 1), rpm=10), duration=6)
}
# With a limited number of levels, viridis can also serve as a discrete
# color scale
attach(iris)
plot(Petal.Length, Sepal.Length, pch=19, col=viridis_map(Species))
legend(1, 8, legend=levels(Species), pch=19,
       col=viridis_colors(n=nlevels(Species)))

}
\seealso{
\code{\link{viridis}} for the colors in the palette.

Other color scales and palettes: \code{\link{brewer_scale}},
  \code{\link{chroma_scale}},
  \code{\link{cubehelix_scale}}, \code{\link{etopo_scale}},
  \code{\link{hue_scale}}, \code{\link{inferno_scale}},
  \code{\link{interp_scale}}, \code{\link{light_scale}},
  \code{\link{magma_scale}}, \code{\link{plasma_scale}},
  \code{\link{wikitopo_scale}}
}
