% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_wikitopo.R
\docType{data}
\name{wikitopo}
\alias{wikitopo}
\title{Colors from Wikipedia topographic conventions}
\format{
A data.frame with 29 rows and 2 variables:
\describe{
  \item{altitude}{the altitude.}
  \item{color}{the hex code of the correpsonding color.}
}
}
\source{
\url{https://en.wikipedia.org/wiki/Wikipedia:WikiProject_Maps/Conventions/Topographic_maps}
}
\usage{
wikitopo
}
\description{
Colors from Wikipedia topographic conventions
}
\seealso{
\code{\link{wikitopo_scale}} to use the palette.
}
\keyword{datasets}
