% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_wikitopo.R
\name{wikitopo_scale}
\alias{wikitopo_scale}
\alias{wikitopo_map}
\alias{wikitopo_palette}
\alias{wikitopo_colors}
\alias{scale_fill_wikitopo}
\alias{scale_color_wikitopo}
\alias{scale_colour_wikitopo}
\title{Wikipedia topographic scale and palette}
\usage{
wikitopo_scale(exact.until = 1000)

wikitopo_map(x, ...)

wikitopo_palette(exact.until = 1000, ...)

wikitopo_colors(n, ...)

scale_fill_wikitopo(...)

scale_color_wikitopo(...)
}
\arguments{
\item{exact.until}{integer, when more than \code{exact.until} colors need to be computed, a fast but not exact alternative algorithm is used. This number is increased compared to the default in \code{\link{interp_scale}} because some transitions in color along the wikipedia topographic palette are sharp, in particular around altitude=0. To get exact interpolation all the time, use a very large number.}

\item{x}{a vector whose values will be coerced to numbers and mapped to colors.}

\item{...}{passed to \code{\link{wikitopo_scale}} or to \code{scale_*_gradientn} for the \code{scale_*_wikitopo} functions.}

\item{n}{number of colors to extract from the color palette.}
}
\value{
\code{*_scale} returns a \strong{function}. This function takes a single argument (\code{x}: a numeric vector), maps its values to colors, and returns thee colors as hex codes.

\code{*_map} is a shortcut for \code{*_scale(domain=range(x))(x)}: it creates a scale that spans the range of values in argument \code{x}, maps the content of \code{x} on that scale, and returns the \strong{colors}.

\code{*_palette} returns a \strong{function}. This function takes an integer (\code{n}) as argument, picks \code{n} colors evenly spaced along the scale, and returns them as hex codes.

\code{*_colors} is a shortcut for \code{*_palette()(n)} and directly returns \code{n} evenly spaced \strong{colors}. It is equivalent to built-in functions such as \code{\link[grDevices]{heat.colors}}, \code{\link[grDevices]{topo.colors}}, etc.

\code{scale_*} return a \strong{ggplot2 scale}, either discrete (similar to \code{\link[ggplot2]{scale_color_discrete}}) or continuous (similar to \code{\link[ggplot2]{scale_color_continuous}}).
}
\description{
Color scale and palette from Wikipedia's topographic maps. These scales are special because the mapping between colors and values is predefined to work for altitudes (or depths when the numbers are negative).
}
\examples{
# Defining a scale in that case has little value since it cannot be
# customized. Explore the default mapping
altitudes <- c(-5000, -1000, -10, 0, 10, 200, 500, 5000)
show_col(wikitopo_map(altitudes))
# Explore the full range of colors
show_col(wikitopo_map(seq(-8000, 8000, length=500)))

# It is possible to get n colors from the palette but they loose their
# association with specific altitudes
wikitopo_colors(5)

# The goal of these scales is to color maps and elevation models. Here is
# one centered on Thailand, showcasing high mountains and deep trenches
levs <- seq(-9000, 6000, by=500)
contour(thai, levels=levs, col=wikitopo_map(levs), asp=1.03)

filled.contour(thai, levels=levs, col=wikitopo_map(levs), asp=1.03)

persp(thai, theta=30, phi=25, border=alpha("black", 0.2), expand=0.2,
      col=wikitopo_map(persp_facets(thai$z)))

\dontrun{
# 3D rotating map
library("rgl")
# To get a perfect land/sea mapping, the colors need to be computed
# exactly based on the input altitude values; set `exact.until` very high
# do achieve this (but this makes the function slower of course).
persp3d(thai, aspect=c(1,0.96,0.2), axes=FALSE, box=FALSE,
        col=wikitopo_map(thai$z, exact.until=10))
play3d(spin3d(axis=c(0, 0, 1), rpm=10), duration=6)

# ggplot2 maps
library("ggplot2")
ggplot(thaixyz) + coord_quickmap() +
  geom_contour(aes(x, y, z=z, colour=..level..), breaks=levs) +
  geom_contour(aes(x, y, z=z), breaks=0, colour="black", linetype="11") +
  theme_light() + scale_colour_wikitopo() +
  scale_xy_map()
ggplot(thaixyz) + coord_quickmap() +
  geom_raster(aes(x, y, fill=z)) +
  scale_fill_wikitopo() +
  geom_contour(aes(x, y, z=z), breaks=0, colour="black", alpha=0.5) +
  scale_xy_map()
}
}
\references{
\url{https://en.wikipedia.org/wiki/Wikipedia:WikiProject_Maps/Conventions/Topographic_maps}.
}
\seealso{
Other color scales and palettes: \code{\link{brewer_scale}},
  \code{\link{chroma_scale}},
  \code{\link{cubehelix_scale}}, \code{\link{etopo_scale}},
  \code{\link{hue_scale}}, \code{\link{inferno_scale}},
  \code{\link{interp_scale}}, \code{\link{light_scale}},
  \code{\link{magma_scale}}, \code{\link{plasma_scale}},
  \code{\link{viridis_scale}}
}
