% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_brewer.R
\docType{data}
\name{brewer}
\alias{brewer}
\title{Colors from colorbrewer2.org}
\format{A data.frame with 1680 rows and 4 variables:
\describe{
  \item{name}{the name of the color palette.}
  \item{type}{the type of palette (\code{qualitative}, \code{diverging}, or \code{sequential}).}
  \item{n}{the number of colors in the palette. For a given palette, all combinations of colors are hand picked and not just interpolated between the two extremes. So, even in the same palette, the colors for the different values of \code{n} may be slightly different.}
  \item{color}{the hex code of the color.}
}}
\source{
\url{http://colorbrewer2.org}
}
\usage{
brewer
}
\description{
Colors from colorbrewer2.org
}
\seealso{
\code{\link{brewer_info}} for a summary of the properties of colorbrewer palettes.
}
\keyword{datasets}

