% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_brewer.R
\docType{data}
\name{brewer_info}
\alias{brewer_info}
\title{Description of the colorbrewer2.org color palettes}
\format{
A data.frame with 35 rows and 3 variables:
\describe{
  \item{type}{the type of palette (\code{qualitative}, \code{diverging}, or \code{sequential}).}
  \item{name}{the name of the color palette.}
  \item{maxcolors}{the maximum number of handpicked colors in the palette. (NB: the minimum is always 3).}
}
}
\source{
\url{http://colorbrewer2.org}
}
\usage{
brewer_info
}
\description{
Description of the colorbrewer2.org color palettes
}
\seealso{
\code{\link{brewer}} for a complete table of colorbrewer2 colors.
}
\keyword{datasets}
