% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel.R
\name{channel}
\alias{channel}
\alias{channel<-}
\title{Get or set color channel}
\usage{
channel(x, model, channel)

channel(x, model, channel) <- value
}
\arguments{
\item{x}{vector of colors specified as hex strings or named R colors.}

\item{model}{string defining the color model; valid models are \code{rgb}, \code{rgba}, \code{hsv}, \code{hsl}, \code{hsi}, \code{hcl}, \code{lch}, \code{lab}, \code{cmyk}, \code{css}, \code{hex}, \code{temperature}.}

\item{channel}{string defining the channel within the color model.}

\item{value}{the channel value to set; a number, the convention of which depends on the channel.}
}
\value{
For \code{channel}, a vector of channel values, the convention of which depend on the channel.

For \code{channel <-}, the updated object(s).
}
\description{
Get or set color channel
}
\examples{
channel("pink", "rgb", "r")
channel(colors()[1:5], "hcl", "l")

# Colors along a HSV rainbow have very different lightness,
# which makes them inapropriate for continuous color scales
channel(rainbow(6), "hcl", "l")
# But, by definition, lightness should be constant along a HCL "rainbow"
channel(hcl(h=seq(0, 360, length.out=6), l=0.5), "hcl", "l")
# NB: subtle changes are caused by the round-trip conversion to R colors.

# Force a given lightness
x_orig <- x <- rainbow(6)
channel(x, "hcl", "l") <- 0.3
show_col(x_orig, x)

# Make all colors equally saturated
x_orig <- x <- c("aliceblue", "aquamarine4", "coral", "blanchedalmond")
channel(x, "hsv", "s") <- 0.5
show_col(x_orig, x)

# Keep the lightness and saturation but change the hue
# (also called "colorizing")

x_orig <- x <- c("aliceblue", "aquamarine4", "coral", "blanchedalmond")
channel(x, "hsv", "h") <- 240  # make all blue
show_col(x_orig, x)
}
\seealso{
\code{\link{luminance}} for relative brightness, which is slightly different from the perceived brightness (i.e. the value of the L channel in L*a*b* or HCL).

Other color manipulation functions: \code{\link{darken}},
  \code{\link{luminance}}, \code{\link{mix}},
  \code{\link{saturate}}
}
