% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmyk.R
\name{cmyk}
\alias{cmyk}
\title{CMYK Color Specification}
\usage{
cmyk(c = 0, m = 0.6, y = 0.6, k = 0.4)
}
\arguments{
\item{c, m, y, k}{color components, numeric vectors with values in \code{[0, 1]}.}
}
\value{
A vector of colors specified as hex codes
}
\description{
Create a vector of colors from cyan, magenta, yellow, and black
}
\details{
The first argument can also be a data.frame or a matrix. In that case, its first three columns are considered as the three color components and the rest of the function arguments are ignored.

When the arguments are used separately and are vectors, values in shorter arguments are recycled to match the length of the longest argument. If the lengths are not compatible, an error is output.
}
\examples{
cmyk()
cmyk(0.47, 0.55, 0.3)
cmyk(data.frame(c(0.4, 0.6), c(-1, 1), c(-1, 1)))
cmyk(matrix(c(0.4, 0.6, -1, 1, -1, 1), ncol=3))
cmyk(0.5, c(-1, 1), c(-1, 1))

# color ramp
show_col(cmyk(k=seq(0, 1, length.out=10)))

}
\seealso{
Other color specifications: \code{\link{color_temperature}},
  \code{\link{css_color}}, \code{\link{hcl}},
  \code{\link{hsl}}, \code{\link{hsv}}, \code{\link{lab}}
}

