% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_color.R
\name{convert_color}
\alias{as.cmyk}
\alias{as.css}
\alias{as.hcl}
\alias{as.hsl}
\alias{as.hsv}
\alias{as.lab}
\alias{as.lch}
\alias{as.rgb}
\alias{convert_color}
\title{Convert colors to a given color model}
\usage{
convert_color(x, model)

as.rgb(x)

as.hsv(x)

as.hsl(x)

as.hcl(x)

as.lch(x)

as.lab(x)

as.cmyk(x)

as.css(x)
}
\arguments{
\item{x}{vector of colors specified as hex strings or named R colors.}

\item{model}{string defining the color model; valid models are \code{rgb}, \code{gl}, \code{hsv}, \code{hsl}, \code{hcl}, \code{lch}, \code{lab}, \code{cmyk}, \code{css}, \code{hex}.}
}
\value{
A matrix containing the color components, except for the \code{css} model for which the result is a vector of css color definition strings.
}
\description{
Convert a vector of R colors to a given color model
}
\examples{
convert_color("red", model="rgb")
convert_color("red", model="hsl")
convert_color("red", model="hcl")
convert_color("red", model="cmyk")

as.lab("red")
as.lch("red")
as.rgb(colors()[1:5])

as.css("red")

}

