% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_color.R
\name{convert_color}
\alias{convert_color}
\alias{as.rgb}
\alias{as.rgba}
\alias{as.hsv}
\alias{as.hsl}
\alias{as.hsi}
\alias{as.hcl}
\alias{as.lch}
\alias{as.lab}
\alias{as.cmyk}
\alias{as.css}
\alias{as.hex}
\alias{as.temperature}
\title{Convert colors to a given color model}
\usage{
convert_color(x, model)

as.rgb(x)

as.rgba(x)

as.hsv(x)

as.hsl(x)

as.hsi(x)

as.hcl(x)

as.lch(x)

as.lab(x)

as.cmyk(x)

as.css(x)

as.hex(x)

as.temperature(x)
}
\arguments{
\item{x}{vector of colors specified as hex strings or named R colors.}

\item{model}{string defining the color model; valid models are \code{rgb}, \code{rgba}, \code{hsv}, \code{hsl}, \code{hsi}, \code{hcl}, \code{lch}, \code{lab}, \code{cmyk}, \code{css}, \code{hex}, \code{temperature}.}
}
\value{
A matrix containing the color components in most cases, except for the models:
\describe{
 \item{\code{css}}{a vector of css color definition strings,}
 \item{\code{hex}}{a vector of hexadecimal strings defining colors,}
 \item{\code{temperature}}{a vector of numbers corresponding to the temperature of the color in Kelvin.}
}
}
\description{
Convert a vector of colors to a given color model.
}
\examples{
convert_color("red", model="rgb")
convert_color("red", model="hsl")
convert_color("red", model="hcl")
convert_color("red", model="cmyk")

as.hsv("red")
as.lab("red")
as.lch("red")

as.css("red")
as.hex("red")
as.temperature("red")

# Can be vectorised
as.rgb(colors()[1:5])
as.rgb(c("#B55FFC", "blue", "purple", "#6A9F16"))

}
\seealso{
See \code{\link{parse_color}} for a general function to parse colors in various specifications.
}
