% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_etopo.R
\name{etopo_scale}
\alias{etopo_scale}
\alias{etopo_map}
\alias{etopo_palette}
\alias{etopo_colors}
\title{ETOPO1 scale and palette}
\usage{
etopo_scale(exact.until = 1000)

etopo_map(x, ...)

etopo_palette(exact.until = 1000, ...)

etopo_colors(n, ...)
}
\arguments{
\item{exact.until}{integer, when more than \code{exact.until} colors need to be computed, a fast but not exact alternative algorithm is used. This number is increased compared to the default in \code{\link{interp_scale}} because some transitions in color along the ETOPO1 palette are sharp, in particular around altitude=0. To get exact interpolation all the time, use a very large number.}

\item{...}{passed to \code{\link{etopo_scale}}.}
}
\value{
\code{*_scale} and \code{*_palette} return a function.

\code{*_map} and \code{*_colors} return a vector of hex colors.
}
\description{
Color scale and palette inspired from the representation of the ETOPO1 global relief model. These scales are special because the mapping between colors and values is predefined to work for altitudes (or depths when the numbers are negative).
}
\details{
As a user, the most useful functions are \code{*_map} and \code{*_colors}.

\code{*_scale} creates a \strong{function} that takes a single argument (\code{x}: a numeric vector), maps its values to colors, and returns those colors as hex codes.

\code{*_map} is a shortcut for \code{*_scale(domain=range(x))(x)}: it creates a scale that spans all values in its argument \code{x}, maps the content of \code{x} on that scale, and returns the colors. It is particularly useful as the \code{col} argument of plotting functions.

\code{*_palette} also creates a \strong{function}, but this one takes an integer (\code{n}) as argument, picks \code{n} colors evenly spaced along the scale, and returns them as hex codes.

\code{*_colors} is a shortcut for \code{*_palette()(n)} and directly returns \code{n} evenly spaced colors. It is equivalent to built-in functions such as \code{\link[grDevices]{heat.colors}}, \code{\link[grDevices]{topo.colors}}, etc.
}
\examples{
# Defining a scale in that case has little value since it cannot be
# customized. Explore the default mapping
altitudes <- c(-5000, -1000, -10, 0, 10, 200, 500, 5000)
show_col(etopo_map(altitudes))
# Explore the full range of colors
show_col(etopo_map(seq(-8000, 8000, length=500)))

# It is possible to get n colors from the palette but they loose their
# association with specific altitudes
etopo_colors(5)

# The goal of these scales is to color maps and elevation models. Here is
# one centered on Thailand, showcasing high mountains and deep trenches
levs <- seq(-9000,6000,by=500)
contour(thai, levels=levs, col=etopo_map(levs), asp=1.03)

filled.contour(thai, levels=levs, col=etopo_map(levs), asp=1.03)

persp(thai, theta=30, phi=25, border=alpha("black", 0.2), expand=0.2,
      col=etopo_map(persp_facets(thai$z)))

\dontrun{
# 3D rotating map
library("rgl")
# To get a perfect land/sea mapping, the colors need to be computed
# exactly based on the input altitude values; set `exact.until` very high
# do achieve this (but this makes the function slower of course).
persp3d(thai, aspect=c(1,0.96,0.2), axes=FALSE, box=FALSE,
        col=etopo_map(thai$z, exact.until=10^5))
play3d(spin3d(axis=c(0, 0, 1), rpm=10), duration=6)
}
}
\references{
The topographic colors are based on GMT globe by Lester M. Anderson of CASP, UK, modified by Jesse Varner and Elliot Lim (NOAA/NGDC) to have a smaller band of white at the highest elevations. The bathymetry is based on GMT haxby, popularised by Bill Haxby, LDEO. See \url{https://www.ngdc.noaa.gov/mgg/global/global.html}.
}
\seealso{
\code{\link{etopo}} for the colors in the palette and the associated altitudes.

Other color scales and palettes: \code{\link{brewer_scale}},
  \code{\link{chroma_scale}},
  \code{\link{cubehelix_scale}}, \code{\link{hue_scale}},
  \code{\link{inferno_scale}}, \code{\link{interp_scale}},
  \code{\link{light_scale}}, \code{\link{magma_scale}},
  \code{\link{plasma_scale}}, \code{\link{viridis_scale}}
}
