% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsl.R
\name{hsl}
\alias{hsl}
\title{HSL Color Specification}
\usage{
hsl(h = 0, s = 0.5, l = 0.5)
}
\arguments{
\item{h}{hue, numeric vector with values in \code{[0, 360]}.}

\item{s}{saturation, numeric vector with values in \code{[0, 1]}; 0 is grey, 1 is full color.}

\item{l}{lightness, numeric vector with values in \code{[0, 1]}; 0 is black, 1 is white.}
}
\value{
A vector of colors specified as hex codes
}
\description{
Create a vector of colors from hue, saturation, and lightness.
}
\details{
The first argument can also be a data.frame or a matrix. In that case, its columns are considered as the color components, taken in order, and the other color components arguments are ignored.

When separate arguments are used for the color components and are vectors, values in shorter arguments are recycled to match the length of the longest argument. If the lengths are not compatible, an error is output.
}
\examples{
hsl()
hsl(330, 1, 0.6)
hsl(data.frame(c(330, 340), c(1,1), c(0.6, 0.4)))
hsl(matrix(c(330, 340, 1, 1, 0.6, 0.4), ncol=3))
hsl(c(330, 340), 1, 0.6)

# color ramps
ramp <- seq(0, 1, length.out=10)
show_col(hsl(h=ramp*360))
show_col(hsl(s=ramp))
show_col(hsl(l=ramp))

}
\seealso{
See \code{\link{parse_color}} for the general function to parse colors in various specifications, which this function calls internally.

Other color specifications: \code{\link{cmyk}},
  \code{\link{color_temperature}}, \code{\link{css_color}},
  \code{\link{hcl}}, \code{\link{hsv}}, \code{\link{lab}},
  \code{\link{parse_color}}, \code{\link{rgb}}
}

