% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_hsv.R
\name{hsv}
\alias{hsv}
\title{HSV Color Specification}
\usage{
hsv(h = 0, s = 0.6, v = 0.7)
}
\arguments{
\item{h}{hue, numeric vector with values in \code{[0, 360]}.}

\item{s}{saturation, numeric vector with values in \code{[0, 1]}; 0 is grey, 1 is full color.}

\item{v}{value, same conventions as \code{s}; 0 is black, 1 is full brightness}
}
\value{
A vector of colors specified as hex codes
}
\description{
Create a vector of colors from hue, saturation, and value.
}
\details{
The first argument can also be a data.frame or a matrix. In that case, its columns are considered as the color components, taken in order, and the other color components arguments are ignored.

When separate arguments are used for the color components and are vectors, values in shorter arguments are recycled to match the length of the longest argument. If the lengths are not compatible, an error is output.
}
\examples{
hsv()
hsv(330, 1, 0.6)
hsv(data.frame(c(330, 340), c(1,1), c(0.6, 0.4)))
hsv(matrix(c(330, 340, 1, 1, 0.6, 0.4), ncol=3))
hsv(c(330, 340), 1, 0.6)

# color ramps
ramp <- seq(0, 1, length.out=10)
show_col(hsv(h=ramp*360))
show_col(hsv(s=ramp))
show_col(hsv(v=ramp))

}
\seealso{
See \code{\link{parse_color}} for the general function to parse colors in various specifications (which this function calls internally) and \code{\link{convert_color}} to convert parsed colors to another model.

Other color specifications: \code{\link{cmyk}},
  \code{\link{color_temperature}}, \code{\link{css_color}},
  \code{\link{hcl}}, \code{\link{hsl}}, \code{\link{lab}},
  \code{\link{parse_color}}, \code{\link{rgb}}
}

