% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luminance.R
\name{luminance}
\alias{luminance}
\alias{luminance<-}
\title{Get or set the relative brightness of colors}
\usage{
luminance(x)

luminance(x) <- value
}
\arguments{
\item{x}{vector of colors specified as hex strings or named R colors.}

\item{value}{the luminance value, as a number in [0,1]}
}
\value{
For \code{luminance}, a vector of luminance values, each in [0,1].

For \code{luminance<-}, the updated object(s).
}
\description{
Get or set the relative brightness of colors
}
\details{
\code{luminance} gets the relative brightness of colors, normalised to 0 for darkest black and 1 for lightest white according to the WCAG definition \url{http://www.w3.org/TR/2008/REC-WCAG20-20081211/#relativeluminancedef}.

\code{luminance<-} sets the brightness by interpolating with black or white until the appropriate relative luminance is found.
}
\examples{
luminance("black")
luminance("white")
luminance("red")

# Colors along a HSV rainbow have very different perceived brightness
luminance(rainbow(6))
# but brightness is much more homogenous along a HCL "rainbow"
luminance(hcl(h=seq(0, 360, length.out=6)))

# Change the luminance
col <- "red"
luminance(col) <- 0.5
show_col(c("red", col))

# This can even be used to create color ramps, although the steps in
# brightness may not be regular at the extremes
col <- "red"
luminance(col) <- seq(0, 1, length.out=10)
show_col(col)
# Better to avoid the extremes then
col <- "red"
luminance(col) <- seq(0.1, 0.9, length.out=10)
show_col(col)

}
\seealso{
\code{\link{channel}} to extract or set a given channel, in particular the lightness channel in HSL, HCL, or L*a*b* color spaces.

Other color manipulation functions: \code{\link{change_lightness}},
  \code{\link{change_saturation}}, \code{\link{channel}},
  \code{\link{mix}}
}

