% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_magma.R
\name{magma_scale}
\alias{magma_scale}
\alias{magma_map}
\alias{magma_palette}
\alias{magma_colors}
\title{Magma color scale and palette}
\usage{
magma_scale(domain = c(0, 1), reverse = FALSE)

magma_map(x, ...)

magma_palette(...)

magma_colors(n, ...)
}
\arguments{
\item{domain}{the values between which the scale is computed.}

\item{reverse}{whether to reverse the order of colors along the scale.}

\item{x}{a vector whose values will be coerced to numbers and mapped to colors.}

\item{...}{passed to \code{\link{magma_scale}}. Note that argument \code{domain} is meaningless in functions other than \code{magma_scale} and passing it through \code{...} is an error.}

\item{n}{number of colors to extract from the color palette.}
}
\value{
\code{*_scale} and \code{*_palette} return a function.

\code{*_map} and \code{*_colors} return a vector of hex colors.
}
\description{
The \code{magma} color palette, by Nathaniel J. Smith and Stefan van der Walt (CC0 license).
}
\details{
As a user, the most useful functions are \code{*_map} and \code{*_colors}.

\code{*_scale} creates a \strong{function} that takes a single argument (\code{x}: a numeric vector), maps its values to colors, and returns those colors as hex codes.

\code{*_map} is a shortcut for \code{*_scale(domain=range(x))(x)}: it creates a scale that spans all values in its argument \code{x}, maps the content of \code{x} on that scale, and returns the colors. It is particularly useful as the \code{col} argument of plotting functions.

\code{*_palette} also creates a \strong{function}, but this one takes an integer (\code{n}) as argument, picks \code{n} colors evenly spaced along the scale, and returns them as hex codes.

\code{*_colors} is a shortcut for \code{*_palette()(n)} and directly returns \code{n} evenly spaced colors. It is equivalent to built-in functions such as \code{\link[grDevices]{heat.colors}}, \code{\link[grDevices]{topo.colors}}, etc.
}
\examples{
# Get a few colors along the palette
show_col(
  magma_palette()(20),
  magma_colors(50),
  magma_colors(20, reverse=TRUE)
)

# Plot the Maunga Whau volcano elevation map
image(maunga, col=magma_colors(100))
contour(maunga, col=alpha("black", 0.2), add=TRUE)

persp(maunga, theta=50, phi=25, border=alpha("black", 0.3),
      col=magma_map(persp_facets(maunga$z)))

\dontrun{library("rgl")
persp3d(maunga, aspect=c(1,0.6,0.3), axes=FALSE, box=FALSE,
        col=magma_map(maunga$z))
play3d(spin3d(axis=c(0, 0, 1), rpm=10), duration=6)
}
# With a limited number of levels, magma can also serve as a discrete
# color scale
attach(iris)
plot(Petal.Length, Sepal.Length, pch=19, col=magma_map(Species))
legend(1, 8, legend=levels(Species), pch=19,
       col=magma_colors(n=nlevels(Species)))

}
\seealso{
\code{\link{magma}} for the colors in the palette.

Other color scales and palettes: \code{\link{brewer_scale}},
  \code{\link{chroma_scale}},
  \code{\link{cubehelix_scale}}, \code{\link{hue_scale}},
  \code{\link{inferno_scale}}, \code{\link{interp_scale}},
  \code{\link{light_scale}}, \code{\link{plasma_scale}},
  \code{\link{viridis_scale}}
}
