% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{mix}
\alias{mix}
\title{Mix two colors}
\usage{
mix(x, y, ratio = 0.5, model = "rgb")
}
\arguments{
\item{x, y}{colors specified as hex strings or named R colors.}

\item{ratio}{amount of y mixed in x; \code{ratio} = 0 means pure \code{x}, \code{ratio} = 1 means pure \code{y}.}

\item{model}{color model to compute the mix in; valid models are \code{rgb}, \code{hsv}, \code{hsl}, \code{hcl}, \code{lab}.}
}
\value{
A vector of colors specified as hex codes
}
\description{
Mix two colors
}
\details{
All arguments can be vectors, as long as they are of compatible length (shorter arguments will be repeated to the length of longer arguments)
}
\examples{
mix("red", "blue")
mix("red", "blue", 0.25)
# Arguments can be vectors and all mixes are returned
mix("red", "blue", c(0.25, 0.5, 0.75))

show_mix <- function(x, y, ...) {show_col(c(x, y, mix(x, y, ...)))}
show_mix("#FEF213", "#146EFD")
show_mix("#FEF213", "#FA000C")
show_mix("red", "blue", ratio=c(0.25, 0.5, 0.75))
show_mix("red", "blue", model=c("rgb", "hsv", "hsl", "hcl", "lab"))

}
\seealso{
Other color manipulation functions: \code{\link{change_lightness}},
  \code{\link{change_saturation}}, \code{\link{channel}},
  \code{\link{luminance}}
}

