% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_color.R
\name{parse_color}
\alias{parse_color}
\title{Parse colors specified in a given model}
\usage{
parse_color(x, model)
}
\arguments{
\item{x}{a matrix or data.frame whose columns specify the color channels or a vector of character string definitions of colors for the \code{css}, \code{hex}, and \code{temperature} color models.}

\item{model}{string defining the color model; valid models are \code{rgb}, \code{rgba}, \code{hsv}, \code{hsl}, \code{hsi}, \code{hcl}, \code{lch}, \code{lab}, \code{cmyk}, \code{css}, \code{hex}, \code{temperature}.}
}
\value{
A vector of colors specified as hex codes
}
\description{
Parse colors specified in a given model
}
\examples{
parse_color(data.frame(h=c(0, 120, 240), s=0.5, v=0.7), model="hsv")
parse_color(data.frame(r=c(255, 0, 0),
                       g=c(0, 255, 0),
                       b=c(0, 0, 255)), model="rgb")
}
\seealso{
\code{\link{convert_color}} to convert parsed colors to another model.

Other color specifications: \code{\link{cmyk}},
  \code{\link{css}}, \code{\link{hcl}}, \code{\link{hex}},
  \code{\link{hsi}}, \code{\link{hsl}}, \code{\link{hsv}},
  \code{\link{lab}}, \code{\link{rgb}},
  \code{\link{temperature}}, \code{\link{wavelength}}
}
