% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_objects.R
\docType{data}
\name{thai}
\alias{thai}
\alias{thaixyz}
\title{Topography around Thailand}
\format{\code{thai} is a list with two coordinate vectors and a matrix of corresponding values, suitable for \code{\link[graphics]{contour}}, \code{\link[graphics]{persp}} etc. \code{thaixyz} contains the same data but formated as a xyz a data.frame with 30351 rows and 3 columns. Both have components:
\describe{
  \item{x,y}{x and y coordinates, in degrees.}
  \item{z}{altitudes, in m.}
}}
\source{
Extracted from NOAA's NGDC database (\url{https://maps.ngdc.noaa.gov/viewers/wcs-client/}) and regridded.
}
\usage{
thai

thaixyz
}
\description{
Topography of the region around Thailand, featuring high mountains and deep trenches, at 0.2° resolution.
}
\keyword{datasets}
