% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_viridis.R
\name{viridis_scale}
\alias{viridis.colors}
\alias{viridis_colors}
\alias{viridis_map}
\alias{viridis_palette}
\alias{viridis_scale}
\title{Viridis color scale and palette}
\usage{
viridis_scale(domain = c(0, 1), reverse = FALSE)

viridis_map(x, ...)

viridis_palette(...)

viridis_colors(n, ...)

viridis.colors(n, ...)
}
\arguments{
\item{domain}{the values between which the scale is computed.}

\item{reverse}{whether to reverse the order of colors along the scale.}

\item{x}{a vector whose values will be coerced to numbers and mapped to colors.}

\item{...}{passed to \code{\link{viridis_scale}}. Note that argument \code{domain} is meaningless in functions other than \code{viridis_scale} and passing it through \code{...} is an error.}

\item{n}{number of colors to extract from the color palette.}
}
\value{
\code{*_scale} and \code{*_palette} return a function.

\code{*_map}, \code{*_colors}, and \code{*.colors} return a vector of hex colors.
}
\description{
The \code{viridis} color palette, by Nathaniel J. Smith, Stefan van der Walt, and Eric Firing (CC0 license).
}
\details{
The \code{*_scale} version creates a \strong{function} that takes a single argument (\code{x}: a numeric vector), maps the values to colors along an interpolated gradient and returns those colors as hex codes.

The \code{*_map} version is a shortcut for \code{*_scale(domain=range(x))(x)}: it creates a scale that spans all values in its argument \code{x}, maps the content of \code{x} on that scale and returns the colors. It is particularly useful as the \code{col} argument of plotting functions.

The \code{*_palette} version also creates a \strong{function}, but this one takes an integer (\code{n}) as argument, picks \code{n} colors evenly spaced along the scale and returns them as hex codes.

The \code{*_colors} and \code{*.colors} versions are shortcuts for \code{*_palette()(n)} and directly return \code{n} evenly spaced colors. They are equivalent to built-in functions such as \code{\link[grDevices]{heat.colors}}, \code{\link[grDevices]{topo.colors}}, etc.
}
\examples{
# Get a few colors along the palette
show_col(
  viridis_palette()(20),
  viridis_colors(50),
  viridis_colors(20, reverse=TRUE)
)

# Plot the Maunga Whau volcano elevation map
x <- 10*(1:nrow(volcano))
y <- 10*(1:ncol(volcano))
image(x, y, volcano, col=viridis_colors(100))
contour(x, y, volcano, col=alpha("black", 0.2), add=TRUE)

persp(x, y, volcano, theta=50, phi=25, border=alpha("black", 0.3),
      col=viridis_map(persp_facets(volcano)))

\dontrun{library("rgl")
persp3d(x, y, volcano, aspect=c(1,0.6,0.3), axes=FALSE, box=FALSE,
        col=viridis_map(volcano))
play3d(spin3d(axis=c(0, 0, 1), rpm=10), duration=6)
}
# With a limited number of levels, viridis can also serve as a discrete
# color scale
attach(iris)
plot(Petal.Length, Sepal.Length, pch=19, col=viridis_map(Species))
legend(1, 8, legend=levels(Species), pch=19,
       col=viridis_colors(n=nlevels(Species)))

}
\seealso{
\code{\link{viridis}} for the colors in the palette.

Other color scales and palettes: \code{\link{brewer_scale}},
  \code{\link{color_scale}}, \code{\link{cubehelix_scale}},
  \code{\link{inferno_scale}}, \code{\link{magma_scale}},
  \code{\link{plasma_scale}}
}

